feat(version7): FIMD Global v7 — Operaciones en horario hábil con colación + pre/post pulso forzados + objetivo de nitrógeno total N0 + (N1+N2) = 450 ppm - Cambios vs v6:   • t=0 ahora es **lunes 12:00** (inoculación a mediodía).   • No se puede **samplear entre 13:00–14:00** ni **después de 17:00** (ventana hábil 12:00–17:00 excl., con pausa).   • Forzado de muestreo **antes** y **después** de cada pulso de N (además de opciones previas).   • Restricción de N: en todos los diseños se ajusta **N1+N2 = 450 − N0** (mg/L).
Mantiene:
  • Operaciones (t12, t23, tN1, tN2) en **horario hábil L–V**.
  • Muestreo ampliado (K_per_exp=18) con **máx. 2 muestras/día** y calendario laboral.
  • Forzados en set-points (activables por flag).
  • Modo de reasignación global de tiempos (más óptimo).
